Set input/output encoding to UTF-8 in the query string.
The query parameters "ie" and "oe" appear to have been introduced only
recently.  When I make a request to translate.google.com using curl,
the response is UTF-8 encoded even if I don't set these parameters.
This is not the case when a request is made from inside Emacs (why?).
Specifying "ie=UTF-8" and "oe=UTF-8" fixes this problem.