Make `google-translate-at-point' produce friendlier error messages.
1. If there is neither active region nor word nearby, `current-word'
returns nil, which `google-translate-translate' can't handle, and it
fails with a cryptic error message about bad types.  We detect this
situation and display a friendlier message "No word at point."

2. Parsing function arguments specified with (interactive "r") fails
with "The mark is not set now, so there is no region" if no mark has
ever been set in a buffer.  This parsing happens even if we want to
translate a word at point.  The error message goes away as soon as
mark is set.  To avoid this nuisance, we use (region-beginning) and
(region-end) only in the branch in which we know that mark is active.