Deal with multi-line, multi-sentence regions.
To deal with multi-line regions, sequences of white space are replaced
by single spaces (thanks to Bernard Hurley for sending a patch).

In cases when the input text consists of multiple lines or sentences,
the returned JSON contains an array of translations of separate lines
resp. sentences.  Previously we used to extract essentially the first
element of that array.  Now we concatenate all translations.  Because
white space has been stripped, to make the output prettier, we fill
(wrap) the translated text.